Address review: widen `maxPolygonToCellsSize` return type to `int64_t`
Returning `i32` truncates `usize` values from `max_cells_count` for large
polygons, which can cause under-allocation and out-of-bounds writes in
`polygonToCells`. Use `int64_t` / `i64` in both the header and Rust FFI,
clamped to `i64::MAX`.

+++ rust/workspace/h3o/include/h3api.h
@@ -142,7 +142,7 @@ int maxFaceCount(H3Index h3);
 void getIcosahedronFaces(H3Index h3, int *out);
 
 /* Polygon */
-int maxPolygonToCellsSize(const GeoPolygon *geoPolygon, int res);
+int64_t maxPolygonToCellsSize(const GeoPolygon *geoPolygon, int res);
 void polygonToCells(const GeoPolygon *geoPolygon, int res, H3Index *out);
 
 #ifdef __cplusplus
